add eks nodepool templating
